[ClavaHls] Uses long instead of int for estimated values

+++ ClavaHls/src/pt/up/fe/specs/clava/hls/strategies/FunctionInlining.java
@@ -35,10 +35,10 @@ import pt.up.fe.specs.clava.hls.directives.HLSInline.InlineType;
 import pt.up.fe.specs.clava.hls.heuristics.InlineHeuristic;
 
 public class FunctionInlining extends RestructuringStrategy {
-    private HashMap<DataFlowNode, Integer> callFreq;
+    private HashMap<DataFlowNode, Long> callFreq;
     private HashMap<String, DataFlowGraph> functions;
     private HashMap<String, Boolean> isFunctionInlinable;
-    private HashMap<String, Integer> functionCosts;
+    private HashMap<String, Long> functionCosts;
     private HashMap<DataFlowNode, Boolean> isCallInlinable;
 
     public FunctionInlining(DataFlowGraph dfg) {
@@ -64,12 +64,12 @@ public class FunctionInlining extends RestructuringStrategy {
 	});
 
 	// Check if each call is inlinable
-	int mainFunCost = analyzeFunction(dfg);
+	long mainFunCost = analyzeFunction(dfg);
 	estimateCallFrequencies(calls);
 
 	for (DataFlowNode call : calls) {
-	    int freq = callFreq.get(call);
-	    int funCost = functionCosts.get(call.getLabel());
+	    long freq = callFreq.get(call);
+	    long funCost = functionCosts.get(call.getLabel());
 
 	    // HEURISTIC HERE
 	    boolean inline = InlineHeuristic.calculate(freq, funCost, mainFunCost);
@@ -117,7 +117,7 @@ public class FunctionInlining extends RestructuringStrategy {
 
     public void printFrequencies() {
 	for (DataFlowNode node : callFreq.keySet()) {
-	    int freq = callFreq.get(node);
+	    long freq = callFreq.get(node);
 	    StringBuilder sb = new StringBuilder();
 	    sb.append("Call to \"").append(node.getLabel()).append("\" (l:")
 		    .append(node.getStmt().getLocation().getStartLine()).append(")");
@@ -129,12 +129,12 @@ public class FunctionInlining extends RestructuringStrategy {
 	}
     }
 
-    public int analyzeFunction(String fun) {
+    public long analyzeFunction(String fun) {
 	DataFlowGraph dfg = functions.get(fun);
 	return analyzeFunction(dfg);
     }
 
-    public int analyzeFunction(DataFlowGraph dfg) {
+    public long analyzeFunction(DataFlowGraph dfg) {
 	ArrayList<DataFlowSubgraphMetrics> metrics = new ArrayList<>();
 	for (DataFlowNode root : dfg.getSubgraphRoots()) {
 	    DataFlowSubgraph sub = dfg.getSubgraph(root);
